Add retries to KCL example test runs to mitigate "websocket closed early"
