+++ rust/kcl-lib/src/simulation_tests.rs
@@ -1,5 +1,4 @@
 use std::{
-    future::Future,
     panic::{AssertUnwindSafe, catch_unwind},
     path::{Path, PathBuf},
 };
@@ -8,9 +7,10 @@ use indexmap::IndexMap;
 
 use crate::{
     ExecOutcome, ExecState, ExecutorContext, ModuleId,
-    errors::{ExecErrorWithState, KclError},
+    errors::KclError,
     exec::KclValue,
     execution::{EnvironmentRef, ModuleArtifactState},
+    test_util::execute_with_retries,
     walk::{Node, walk},
 };
 #[cfg(feature = "artifact-graph")]
@@ -39,9 +39,6 @@ struct Test {
 }
 
 pub(crate) const RENDERED_MODEL_NAME: &str = "rendered_model.png";
-/// Additional attempts after the initial execution when the engine yields a
-/// transient error.
-const EXECUTE_RETRIES_FROM_ENGINE_ERROR: usize = 2;
 
 #[cfg(feature = "artifact-graph")]
 const REPO_ROOT: &str = "../..";
@@ -252,35 +249,6 @@ async fn unparse_test(test: &Test) {
         future.await.unwrap().unwrap();
     }
     input_result.unwrap();
-}
-
-/// If execution results in `EngineHangup` or `EngineInternal`, retry.
-async fn execute_with_retries<F, Fut, T>(mut execute: F) -> Result<T, ExecErrorWithState>
-where
-    F: FnMut() -> Fut,
-    Fut: Future<Output = Result<T, ExecErrorWithState>>,
-{
-    let mut retries_remaining = EXECUTE_RETRIES_FROM_ENGINE_ERROR;
-    loop {
-        // Run the closure to execute.
-        let exec_result = execute().await;
-
-        if retries_remaining > 0
-            && let Err(error) = &exec_result
-            && let crate::errors::ExecError::Kcl(kcl_error) = &error.error
-            && matches!(
-                &kcl_error.error,
-                KclError::EngineHangup { .. } | KclError::EngineInternal { .. }
-            )
-        {
-            let error_type = kcl_error.error.error_type();
-            eprintln!("Execute got {error_type}; retrying...");
-            retries_remaining -= 1;
-            continue;
-        }
-
-        return exec_result;
-    }
 }
 
 async fn execute(test_name: &str, render_to_png: bool) {
